backend/requirements.txt
Django>=5.2
django-cors-headers
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
gunicorn
celery[redis]
redis
